fix(brew-notes): display accurate durations

+++ src/main/kotlin/dev/jsinco/recipes/recipe/process/steps/AgeStep.kt
@@ -12,7 +12,7 @@ class AgeStep(val agingTicks: Long, val barrelType: BarrelType) : Step {
     override fun getType(): StepType = StepType.AGE
     override fun display(): Component = Component.translatable(
         "gui.recipes.lore.step.age",
-        Argument.tagResolver(Formatter.number("aging_years", agingTicks / agingYearTicks))
+        Argument.tagResolver(Formatter.number("aging_years", agingTicks.toDouble() / agingYearTicks))
     )
 
     enum class BarrelType {
